requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
openpyxl==3.1.2
python-dotenv==1.0.0

